dev-unique-constraint
* Enforce unique constraint on SN column
* Skip unique rows during insert
* Move sql transactions to sql_ops file
* Change Apache License to MIT
